+++ src/api/withdrawOnchain.test.js
@@ -1,17 +1,22 @@
+const nock = require('nock')
 const instance = require('./test/helpers/instance')
-const mockGasPrice = require('./test/fixtures/getSafeGasPrice')
 const mockGetConf = require('./test/fixtures/getConf')
 
 let dvf
 
 describe('dvf.withdrawOnchain', () => {
   beforeAll(async () => {
+    nock.cleanAll()
     mockGetConf()
-    mockGasPrice()
     dvf = await instance()
   })
 
+  beforeEach(() => {
+    nock.cleanAll()
+  })
+
   it(`Withdraw ETH or ERC20 from onchain call to stark ex`, async () => {
+    mockGetConf()
     const token = 'ETH'
 
     const response = await dvf.withdrawOnchain(token)