+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.5 - 2020-07-02
+### Added
+- Use dvf pub api to get range of gas prices
+
+## 1.0.4 - 2020-07-01
+### Added
+- gas Station API optional
+- Use recommended gas price from config
 
 ## 1.0.3 - 2020-06-26
 ### Added